+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 #### rumqttc
 - **changed** Don't cancel the eventloop when `Client` is dropped. 
+- **changed** Remove network from public interface [**breaking**]
+- **internal** Fuse io into state
+- **internal** Improve packet framing
 
 #### rumqttd
 - **fixed** 0 keepalive handling